+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/ui/MainActivity.kt
@@ -12,6 +12,10 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.appcompat.widget.Toolbar
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.core.view.isVisible
 import androidx.core.view.updatePadding
 import androidx.fragment.app.Fragment
@@ -23,6 +27,7 @@ import androidx.transition.Slide
 import au.com.shiftyjelly.pocketcasts.R
 import au.com.shiftyjelly.pocketcasts.account.AccountActivity
 import au.com.shiftyjelly.pocketcasts.account.PromoCodeUpgradedFragment
+import au.com.shiftyjelly.pocketcasts.account.onboarding.OnboardingFlowComposable
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.analytics.FirebaseAnalyticsTracker
@@ -120,6 +125,7 @@ import au.com.shiftyjelly.pocketcasts.views.R as VR
 import com.google.android.material.R as MR
 
 private const val SAVEDSTATE_PLAYER_OPEN = "player_open"
+private const val SAVEDSTATE_MINIPLAYER_SHOWN = "miniplayer_shown"
 
 @AndroidEntryPoint
 class MainActivity :
@@ -219,7 +225,10 @@ class MainActivity :
             activity = this
         )
 
-        setupPlayerViews()
+        val showMiniPlayerImmediately = savedInstanceState?.getBoolean(SAVEDSTATE_MINIPLAYER_SHOWN, false) ?: false
+        binding.playerBottomSheet.isVisible = showMiniPlayerImmediately
+
+        setupPlayerViews(showMiniPlayerImmediately)
 
         if (savedInstanceState != null) {
             val videoComingToPortrait =
@@ -264,6 +273,18 @@ class MainActivity :
 
         if (BuildConfig.END_OF_YEAR_ENABLED) {
             setupEndOfYearLaunchBottomSheet()
+        }
+
+        binding.onboardingFrame.setContent {
+            if (BuildConfig.ONBOARDING_ENABLED) {
+                var show by rememberSaveable { mutableStateOf(true) }
+                if (show) {
+                    OnboardingFlowComposable(
+                        activeTheme = theme.activeTheme,
+                        completeOnboarding = { show = false },
+                    )
+                }
+            }
         }
     }
 
@@ -313,6 +334,7 @@ class MainActivity :
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
         outState.putBoolean(SAVEDSTATE_PLAYER_OPEN, binding.playerBottomSheet.isPlayerOpen)
+        outState.putBoolean(SAVEDSTATE_MINIPLAYER_SHOWN, binding.playerBottomSheet.isShown)
     }
 
     override fun overrideNextRefreshTimer() {
@@ -477,7 +499,7 @@ class MainActivity :
 
     @OptIn(DelicateCoroutinesApi::class)
     @Suppress("DEPRECATION")
-    private fun setupPlayerViews() {
+    private fun setupPlayerViews(showMiniPlayerImmediately: Boolean) {
         binding.playerBottomSheet.listener = this
 
         viewModel.playbackState.observe(this) { state ->
@@ -515,7 +537,11 @@ class MainActivity :
                 .toFlowable(BackpressureStrategy.LATEST)
         observeUpNext = LiveDataReactiveStreams.fromPublisher(upNextQueueObservable)
         observeUpNext.observe(this) { upNext ->
-            binding.playerBottomSheet.setUpNext(upNext, theme)
+            binding.playerBottomSheet.setUpNext(
+                upNext = upNext,
+                theme = theme,
+                shouldAnimateOnAttach = !showMiniPlayerImmediately
+            )
         }
 
         viewModel.signInState.observe(this) { signinState ->