+++ modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/StoriesPage.kt
@@ -1,5 +1,6 @@
 package au.com.shiftyjelly.pocketcasts.endofyear
 
+import android.graphics.Bitmap
 import androidx.annotation.FloatRange
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
@@ -41,11 +42,20 @@ import au.com.shiftyjelly.pocketcasts.compose.components.TextP50
 import au.com.shiftyjelly.pocketcasts.compose.preview.ThemePreviewParameterProvider
 import au.com.shiftyjelly.pocketcasts.endofyear.StoriesViewModel.State
 import au.com.shiftyjelly.pocketcasts.endofyear.stories.Story
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryFake1
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryFake2
-import au.com.shiftyjelly.pocketcasts.endofyear.views.StoryFake1View
-import au.com.shiftyjelly.pocketcasts.endofyear.views.StoryFake2View
-import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
+import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryListenedCategories
+import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryListenedNumbers
+import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryListeningTime
+import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryTopFivePodcasts
+import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryTopListenedCategories
+import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryTopPodcast
+import au.com.shiftyjelly.pocketcasts.endofyear.views.convertibleToBitmap
+import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryListenedCategoriesView
+import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryListenedNumbersView
+import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryListeningTimeView
+import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryTopFivePodcastsView
+import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryTopListenedCategoriesView
+import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryTopPodcastView
+import au.com.shiftyjelly.pocketcasts.models.db.helper.ListenedNumbers
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
@@ -53,9 +63,10 @@ private val ShareButtonStrokeWidth = 2.dp
 private val StoryViewCornerSize = 10.dp
 
 @Composable
-fun StoriesScreen(
+fun StoriesPage(
     viewModel: StoriesViewModel,
     onCloseClicked: () -> Unit,
+    onShareClicked: (() -> Bitmap) -> Unit,
 ) {
     val state: State by viewModel.state.collectAsState()
     val progress: Float by viewModel.progress.collectAsState()
@@ -67,8 +78,10 @@ fun StoriesScreen(
             onSkipNext = { viewModel.skipNext() },
             onPause = { viewModel.pause() },
             onStart = { viewModel.start() },
-            onCloseClicked = onCloseClicked
+            onCloseClicked = onCloseClicked,
+            onShareClicked = onShareClicked,
         )
+
         State.Loading -> StoriesLoadingView(onCloseClicked)
         State.Error -> StoriesErrorView(onCloseClicked)
     }
@@ -83,81 +96,73 @@ private fun StoriesView(
     onPause: () -> Unit,
     onStart: () -> Unit,
     onCloseClicked: () -> Unit,
+    onShareClicked: (() -> Bitmap) -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    var screenWidth by remember { mutableStateOf(1) }
-    var isPaused by remember { mutableStateOf(false) }
+    Column(modifier = modifier.background(color = Color.Black)) {
+        var onCaptureBitmap: (() -> Bitmap)? = null
+        state.currentStory?.let { story ->
+            Box(modifier = modifier.weight(weight = 1f, fill = true)) {
+                if (!story.isInteractive) {
+                    onCaptureBitmap =
+                        convertibleToBitmap(content = { StorySharableContent(story, modifier) })
+                }
+                StorySwitcher(
+                    onSkipPrevious = onSkipPrevious,
+                    onSkipNext = onSkipNext,
+                    onPause = onPause,
+                    onStart = onStart,
+                ) {
+                    if (story.isInteractive) {
+                        onCaptureBitmap =
+                            convertibleToBitmap(content = { StorySharableContent(story, modifier) })
+                    }
+                }
+                SegmentedProgressIndicator(
+                    progress = progress,
+                    segmentsData = state.segmentsData,
+                    modifier = modifier
+                        .padding(8.dp)
+                        .fillMaxWidth(),
+                )
+                CloseButtonView(onCloseClicked)
+            }
+        }
+        requireNotNull(onCaptureBitmap).let {
+            ShareButton(
+                onClick = { onShareClicked.invoke(it) }
+            )
+        }
+    }
+}
+
+@Composable
+private fun StorySharableContent(
+    story: Story,
+    modifier: Modifier,
+) {
     Box(
         modifier = modifier
             .fillMaxSize()
-            .background(color = Color.Black)
-            .onGloballyPositioned {
-                screenWidth = it.size.width
-            }
-            .pointerInput(Unit) {
-                detectTapGestures(
-                    onTap = {
-                        if (!isPaused) {
-                            if (it.x > screenWidth / 2) {
-                                onSkipNext()
-                            } else {
-                                onSkipPrevious()
-                            }
-                        }
-                    },
-                    onLongPress = {
-                        isPaused = true
-                        onPause()
-                    },
-                    onPress = {
-                        awaitRelease()
-                        if (isPaused) {
-                            onStart()
-                            isPaused = false
-                        }
-                    }
-                )
-            }
+            .clip(RoundedCornerShape(StoryViewCornerSize))
+            .background(color = story.backgroundColor),
+        contentAlignment = Alignment.Center
     ) {
-        state.currentStory?.let {
-            StoryView(it)
+        when (story) {
+            is StoryListeningTime -> StoryListeningTimeView(story)
+            is StoryListenedCategories -> StoryListenedCategoriesView(story)
+            is StoryTopListenedCategories -> StoryTopListenedCategoriesView(story)
+            is StoryListenedNumbers -> StoryListenedNumbersView(story)
+            is StoryTopPodcast -> StoryTopPodcastView(story)
+            is StoryTopFivePodcasts -> StoryTopFivePodcastsView(story)
         }
-        SegmentedProgressIndicator(
-            progress = progress,
-            segmentsData = state.segmentsData,
-            modifier = modifier
-                .padding(8.dp)
-                .fillMaxWidth(),
-        )
-        CloseButtonView(onCloseClicked)
     }
 }
 
 @Composable
-private fun StoryView(
-    story: Story,
-    modifier: Modifier = Modifier,
+private fun ShareButton(
+    onClick: () -> Unit,
 ) {
-    Column {
-        Box(
-            modifier = modifier
-                .fillMaxSize()
-                .weight(weight = 1f, fill = true)
-                .clip(RoundedCornerShape(StoryViewCornerSize))
-                .background(color = story.backgroundColor),
-            contentAlignment = Alignment.Center
-        ) {
-            when (story) {
-                is StoryFake1 -> StoryFake1View(story)
-                is StoryFake2 -> StoryFake2View(story)
-            }
-        }
-        ShareButton()
-    }
-}
-
-@Composable
-private fun ShareButton() {
     RowOutlinedButton(
         text = stringResource(id = LR.string.share),
         border = BorderStroke(ShareButtonStrokeWidth, Color.White),
@@ -167,7 +172,9 @@ private fun ShareButton() {
                 contentColor = Color.White,
             ),
         iconImage = Icons.Default.Share,
-        onClick = {}
+        onClick = {
+            onClick.invoke()
+        }
     )
 }
 
@@ -244,6 +251,52 @@ private fun StoriesEmptyView(
     }
 }
 
+@Composable
+private fun StorySwitcher(
+    onSkipPrevious: () -> Unit,
+    onSkipNext: () -> Unit,
+    onPause: () -> Unit,
+    onStart: () -> Unit,
+    modifier: Modifier = Modifier,
+    content: (@Composable () -> Unit)?,
+) {
+    var screenWidth by remember { mutableStateOf(1) }
+    var isPaused by remember { mutableStateOf(false) }
+    Box(
+        modifier = modifier
+            .fillMaxSize()
+            .onGloballyPositioned {
+                screenWidth = it.size.width
+            }
+            .pointerInput(Unit) {
+                detectTapGestures(
+                    onTap = {
+                        if (!isPaused) {
+                            if (it.x > screenWidth / 2) {
+                                onSkipNext()
+                            } else {
+                                onSkipPrevious()
+                            }
+                        }
+                    },
+                    onLongPress = {
+                        isPaused = true
+                        onPause()
+                    },
+                    onPress = {
+                        awaitRelease()
+                        if (isPaused) {
+                            onStart()
+                            isPaused = false
+                        }
+                    }
+                )
+            }
+    ) {
+        content?.invoke()
+    }
+}
+
 @Preview(showBackground = true)
 @Composable
 private fun StoriesScreenPreview(
@@ -252,7 +305,7 @@ private fun StoriesScreenPreview(
     AppTheme(themeType) {
         StoriesView(
             state = State.Loaded(
-                currentStory = StoryFake1(listOf(Podcast())),
+                currentStory = StoryListenedNumbers(ListenedNumbers(numberOfEpisodes = 1, numberOfPodcasts = 1)),
                 segmentsData = State.Loaded.SegmentsData(
                     xStartOffsets = listOf(0.0f, 0.28f),
                     widths = listOf(0.25f, 0.75f)
@@ -263,7 +316,8 @@ private fun StoriesScreenPreview(
             onSkipNext = {},
             onPause = {},
             onStart = {},
-            onCloseClicked = {}
+            onCloseClicked = {},
+            onShareClicked = {}
         )
     }
 }