+++ modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/StoriesViewModel.kt
@@ -1,15 +1,20 @@
 package au.com.shiftyjelly.pocketcasts.endofyear
 
+import android.content.Context
+import android.graphics.Bitmap
 import androidx.annotation.FloatRange
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import au.com.shiftyjelly.pocketcasts.endofyear.StoriesViewModel.State.Loaded.SegmentsData
 import au.com.shiftyjelly.pocketcasts.endofyear.stories.Story
+import au.com.shiftyjelly.pocketcasts.utils.FileUtilWrapper
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import timber.log.Timber
+import java.io.File
 import java.util.Timer
 import javax.inject.Inject
 import kotlin.concurrent.fixedRateTimer
@@ -18,6 +23,7 @@ import kotlin.math.roundToInt
 @HiltViewModel
 class StoriesViewModel @Inject constructor(
     private val storiesDataSource: StoriesDataSource,
+    private val fileUtilWrapper: FileUtilWrapper,
 ) : ViewModel() {
     private val mutableState = MutableStateFlow<State>(State.Loading)
     val state: StateFlow<State> = mutableState
@@ -25,30 +31,39 @@ class StoriesViewModel @Inject constructor(
     private val mutableProgress = MutableStateFlow(0f)
     val progress: StateFlow<Float> = mutableProgress
 
+    private val stories = MutableStateFlow(emptyList<Story>())
     private val numOfStories: Int
-        get() = storiesDataSource.numOfStories
+        get() = stories.value.size
 
     private var currentIndex: Int = 0
     private val nextIndex
         get() = (currentIndex.plus(1)).coerceAtMost(numOfStories.minus(1))
+    private val totalLengthInMs
+        get() = storyLengthsInMs.sum() + gapLengthsInMs
+    private val storyLengthsInMs: List<Long>
+        get() = stories.value.map { it.storyLength }
+    private val gapLengthsInMs: Long
+        get() = STORY_GAP_LENGTH_MS * numOfStories.minus(1).coerceAtLeast(0)
 
     private var timer: Timer? = null
 
     init {
         viewModelScope.launch {
-            storiesDataSource.loadStories().collect { stories ->
-                val state = if (stories.isEmpty()) {
+            storiesDataSource.loadStories().stateIn(viewModelScope).collect { result ->
+                cancelTimer()
+                if (result.size != stories.value.size) resetProgressAndCurrentIndex()
+                stories.value = result
+
+                val state = if (result.isEmpty()) {
                     State.Error
                 } else {
-                    with(storiesDataSource) {
-                        State.Loaded(
-                            currentStory = storyAt(currentIndex),
-                            segmentsData = SegmentsData(
-                                xStartOffsets = List(numOfStories) { getXStartOffsetAtIndex(it) },
-                                widths = storyLengthsInMs.map { it / totalLengthInMs.toFloat() },
-                            )
+                    State.Loaded(
+                        currentStory = result[currentIndex],
+                        segmentsData = SegmentsData(
+                            xStartOffsets = List(numOfStories) { getXStartOffsetAtIndex(it) },
+                            widths = storyLengthsInMs.map { it / totalLengthInMs.toFloat() },
                         )
-                    }
+                    )
                 }
                 mutableState.value = state
                 if (state is State.Loaded) start()
@@ -59,7 +74,7 @@ class StoriesViewModel @Inject constructor(
     fun start() {
         val currentState = state.value as State.Loaded
         val progressFraction =
-            (PROGRESS_UPDATE_INTERVAL_MS / storiesDataSource.totalLengthInMs.toFloat())
+            (PROGRESS_UPDATE_INTERVAL_MS / totalLengthInMs.toFloat())
                 .coerceAtMost(PROGRESS_END_VALUE)
 
         timer = fixedRateTimer(period = PROGRESS_UPDATE_INTERVAL_MS) {
@@ -69,7 +84,7 @@ class StoriesViewModel @Inject constructor(
             if (newProgress.roundOff() == getXStartOffsetAtIndex(nextIndex).roundOff()) {
                 currentIndex = nextIndex
                 mutableState.value =
-                    currentState.copy(currentStory = storiesDataSource.storyAt(currentIndex))
+                    currentState.copy(currentStory = stories.value[currentIndex])
             }
 
             mutableProgress.value = newProgress
@@ -95,12 +110,17 @@ class StoriesViewModel @Inject constructor(
         mutableProgress.value = getXStartOffsetAtIndex(index)
         currentIndex = index
         mutableState.value =
-            (state.value as State.Loaded).copy(currentStory = storiesDataSource.storyAt(index))
+            (state.value as State.Loaded).copy(currentStory = stories.value[index])
     }
 
     private fun cancelTimer() {
         timer?.cancel()
         timer = null
+    }
+
+    private fun resetProgressAndCurrentIndex() {
+        mutableProgress.value = 0f
+        currentIndex = 0
     }
 
     override fun onCleared() {
@@ -113,12 +133,29 @@ class StoriesViewModel @Inject constructor(
     @FloatRange(from = 0.0, to = 1.0)
     fun getXStartOffsetAtIndex(index: Int): Float {
         val sumOfStoryLengthsTillIndex = try {
-            storiesDataSource.storyLengthsInMs.subList(0, index).sum()
+            storyLengthsInMs.subList(0, index).sum()
         } catch (e: IndexOutOfBoundsException) {
             Timber.e("Story offset checked at invalid index")
             0L
         }
-        return (sumOfStoryLengthsTillIndex + StoriesDataSource.STORY_GAP_LENGTH_MS * index) / storiesDataSource.totalLengthInMs.toFloat()
+        return (sumOfStoryLengthsTillIndex + STORY_GAP_LENGTH_MS * index) / totalLengthInMs.toFloat()
+    }
+
+    fun onShareClicked(
+        onCaptureBitmap: () -> Bitmap,
+        context: Context,
+        showShareForFile: (File) -> Unit
+    ) {
+        pause()
+        viewModelScope.launch {
+            val savedFile = fileUtilWrapper.saveBitmapToFile(
+                onCaptureBitmap.invoke(),
+                context,
+                EOY_STORY_SAVE_FOLDER_NAME,
+                EOY_STORY_SAVE_FILE_NAME
+            )
+            savedFile?.let { showShareForFile.invoke(it) }
+        }
     }
 
     sealed class State {
@@ -132,12 +169,16 @@ class StoriesViewModel @Inject constructor(
                 val xStartOffsets: List<Float> = emptyList(),
             )
         }
+
         object Error : State()
     }
 
     companion object {
+        private const val STORY_GAP_LENGTH_MS = 100L
         private const val PROGRESS_START_VALUE = 0f
         private const val PROGRESS_END_VALUE = 1f
         private const val PROGRESS_UPDATE_INTERVAL_MS = 10L
+        private const val EOY_STORY_SAVE_FOLDER_NAME = "eoy_images_cache"
+        private const val EOY_STORY_SAVE_FILE_NAME = "eoy_shared_image.png"
     }
 }