+++ modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/stories/EndOfYearStoriesDataSource.kt
@@ -1,31 +1,43 @@
 package au.com.shiftyjelly.pocketcasts.endofyear.stories
 
 import au.com.shiftyjelly.pocketcasts.endofyear.StoriesDataSource
-import au.com.shiftyjelly.pocketcasts.repositories.podcast.EndOfYearManager
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.EndOfYearManager
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
-import timber.log.Timber
 import javax.inject.Inject
 
 class EndOfYearStoriesDataSource @Inject constructor(
     private val endOfYearManager: EndOfYearManager,
-) : StoriesDataSource() {
-    override val stories = mutableListOf<Story>()
+) : StoriesDataSource {
+    override suspend fun loadStories(): Flow<List<Story>> {
+        return combine(
+            endOfYearManager.getTotalListeningTimeInSecsForYear(YEAR),
+            endOfYearManager.findListenedCategoriesForYear(YEAR),
+            endOfYearManager.findListenedNumbersForYear(YEAR),
+            endOfYearManager.findTopPodcastsForYear(YEAR, limit = 5),
+        ) { listeningTime, listenedCategories, listenedNumbers, topFivePodcasts ->
+            val stories = mutableListOf<Story>()
 
-    override suspend fun loadStories() =
-        combine(
-            endOfYearManager.findRandomPodcasts(),
-            endOfYearManager.findRandomEpisode()
-        ) { podcasts, episode ->
-            if (podcasts.isNotEmpty()) stories.add(StoryFake1(podcasts))
-            episode?.let { stories.add(StoryFake2(it)) }
+            listeningTime?.let { stories.add(StoryListeningTime(it)) }
+            if (listenedCategories.isNotEmpty()) {
+                stories.add(StoryListenedCategories(listenedCategories))
+                stories.add(StoryTopListenedCategories(listenedCategories))
+            }
+            if (listenedNumbers.numberOfEpisodes > 1 && listenedNumbers.numberOfPodcasts > 1) {
+                stories.add(StoryListenedNumbers(listenedNumbers))
+            }
+            if (topFivePodcasts.isNotEmpty()) {
+                stories.add(StoryTopPodcast(topFivePodcasts.first()))
+                if (topFivePodcasts.size > 1) {
+                    stories.add(StoryTopFivePodcasts(topFivePodcasts))
+                }
+            }
 
             stories
         }
+    }
 
-    override fun storyAt(index: Int) = try {
-        stories[index]
-    } catch (e: IndexOutOfBoundsException) {
-        Timber.e("Story index is out of bounds")
-        null
+    companion object {
+        private const val YEAR = 2022
     }
 }