+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/components/PodcastItem.kt
@@ -1,7 +1,6 @@
 package au.com.shiftyjelly.pocketcasts.compose.components
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -15,11 +14,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import au.com.shiftyjelly.pocketcasts.compose.theme
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
+
+private val PodcastItemIconSize = 64.dp
 
 @Composable
 fun PodcastItem(
@@ -27,6 +29,7 @@ fun PodcastItem(
     onClick: (() -> Unit)?,
     modifier: Modifier = Modifier,
     tintColor: Color? = null,
+    iconSize: Dp = PodcastItemIconSize,
     subscribed: Boolean = false,
     showSubscribed: Boolean = false,
     showDivider: Boolean = true,
@@ -39,14 +42,13 @@ fun PodcastItem(
                 .padding(horizontal = 16.dp)
                 .then(if (onClick == null) Modifier else Modifier.clickable { onClick() })
         ) {
-            Box(modifier = Modifier.padding(top = 4.dp, end = 12.dp, bottom = 4.dp)) {
-                PodcastImage(
-                    uuid = podcast.uuid,
-                    modifier = Modifier.size(64.dp)
-                )
-            }
+            PodcastImage(
+                uuid = podcast.uuid,
+                modifier = modifier.size(iconSize)
+                    .padding(top = 4.dp, end = 12.dp, bottom = 4.dp)
+            )
             Column(
-                modifier = Modifier
+                modifier = modifier
                     .padding(end = 16.dp)
                     .weight(1f)
             ) {